seo_blog_generator/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
google-generativeai==0.8.3
supabase==2.8.1 